@@ -1,17 +1,15 @@
 {
   "name": "index",
   "version": "1.0.0",
-  "description": "description",
+  "description": "your description here",
   "main": "src/index.js",
   "directories": {
     "doc": "docs",
-    "lib": "lib",
     "test": "test"
   },
   "scripts": {
-    "start": "",
-    "build": "npx gulp default",
-    "test": "npm run build && npm run lint && node test/index.spec.js",
+    "build": "esbuild src/index.js --bundle --minify --sourcemap --outfile=dist/bundle.min.js",
+    "test": "npm run lint && npm run build && deno run test/index.spec.js",
     "lint": "npx eslint src/index.js"
   },
   "keywords": [
@@ -19,14 +17,8 @@
   ],
   "author": "Geoff Doty",
   "license": "MIT",
+  "dependencies": {},
   "devDependencies": {
-    "eslint": "^7.9.0",
-    "tape": "^5.0.1"
-  },
-  "dependencies": {
-    "gulp": "^4.0.2",
-    "gulp-minify": "^3.1.0",
-    "gulp-rollup-2": "^1.1.0",
-    "gulp-strip-comments": "^2.5.2"
+    "eslint": "^7.16.0"
   }
 }
